balance progress

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public class Account {
     private int account_id;
     private int user_id;
-    private BigDecimal balance;
+    private double balance;
 
     public int getAccount_id() {
         return account_id;
@@ -23,11 +23,11 @@ public class Account {
         this.user_id = user_id;
     }
 
-    public BigDecimal getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
 }
